+++ FiguresBasic/Point.cpp
@@ -1,8 +1,8 @@
 #include "Point.h"
 
 double Point::dist(Point other) {
-    double xd = xval - other.xval;
-    double yd = yval - other.yval;
+    int xd = xval - other.xval;
+    int yd = yval - other.yval;
     return sqrt(xd*xd + yd * yd);
 }
